requirements.txt
google_api_python_client
langchain
openai
python-dotenv
tiktoken
tqdm
youtube_transcript_api

